Aplicando Resources e boas praticas no Service com autenticacao. (#6)
* ✨ refactor: aplicar UserResource e injetar Service

* ✨ refactor:aplicando Resource, Enums e autenticacao em Address

* ✨ refactor:aplicando Resource, Enums e autenticacao em Category

* ✨ refactor:aplicando Resource, Enums e autenticacao em Coupon

* ✨ refactor:aplicando Resource, Enums e autenticacao em Discount

* ✨ refactor:aplicando Resource, Enums e autenticacao em Auth

* ✨ refactor:aplicando Resource, Enums e autenticacao em Product

* ✨ refactor:aplicando Resource, Enums e autenticacao em User

* ✨ refactor:aplicando Resource, Enums e autenticacao em Cart

* ✨ refactor:aplicando Resource, Enums e autenticacao em CartItem

* ✨ refactor:aplicando Resource, Enums e autenticacao em Order

* ✨ refactor: mudando alguns erros depois de testes

* ✨ refactor: mudando alguns erros depois de testes  2.0

+++ app/Http/Services/Moderator/ProductService.php
@@ -2,8 +2,10 @@
 
 namespace App\Http\Services\Moderator;
 
+use App\Enums\Moderator\ProductDeleteStatus;
 use App\Http\Repository\Moderator\ProductRepository;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Illuminate\Http\Request;
 
 class ProductService
 {
@@ -12,7 +14,7 @@ class ProductService
 
     public function getAllProducts()
     {
-        return $this->productRepository->All();
+        return $this->productRepository->all();
     }
 
     public function findProductById(string $id)
@@ -22,11 +24,21 @@ class ProductService
         } catch (ModelNotFoundException $e) {
             return null;
         }
-    }
+    }   
 
     public function createProduct(array $dataValidated)
     {
         return $this->productRepository->create($dataValidated);
+    }
+
+    public function uploadImage(Request $request)
+    {
+        if($request->hasFile('image'))
+        {
+            return $request->file('image')->store('products', 'public');
+        }
+
+        return null;
     }
 
     public function updateProduct(array $dataValidated, string $id)
@@ -49,12 +61,16 @@ class ProductService
 
         if(!$product)
         {
-            return null;
+            return ProductDeleteStatus::NOT_FOUND;
+        }   
+        if ($product->orderItems()->exists())
+        {
+            return ProductDeleteStatus::HAS_ORDERS_ITEMS;
         }
-        
+
         $product->cartItems()->delete();
         $product->delete();
 
-        return $product;
+        return ProductDeleteStatus::DELETED;
     }
 }
